// src/enums/screenNames.ts
export enum screenNames {
  HOME = "HOME",
  LOGIN = "LOGIN",
  REGISTER = "REGISTER",
  DRAWERNAVIGATION = "DRAWERNAVIGATION",
  BOTTOMNAVIGATION = "BOTTOMNAVIGATION",
  PHONE = "PHONE",
  MESSAGE = "MESSAGE",
  OTPSCREEN = "OTPSCREEN",
  ALLCHATMESSAGES = "ALLCHATMESSAGES",
  PROFILE = "PROFILE",
  EDITPROFILE = "EDITPROFILE",
  EPISODEDETAILS = "EPISODEDETAILS",
  TOCINFORMATION = "TOCINFORMATION",
  NOTIFICATIONS = "NOTIFICATIONS",
  SPLASHSCREEN = "SPLASHSCREEN",
  PRESURGERY = "PRESURGERY",
  CHAT = "CHAT",
  CONFIRMPASSWORD = "CONFIRMPASSWORD",
  RESETPASSWORD = "RESETPASSWORD",
  PREREGISTER = "PREREGISTER",
  CREATECONVERSION = "CREATECONVERSION",
  PDFVIEWERSCREEN = "PDFVIEWERSCREEN",
  ABOUTSCREEN = "ABOUTSCREEN",
  FACEIDTOUCHIDENABLE = "FACEIDTOUCHIDENABLE",
}